refactor: src/shared/utils에서 src/shared/lib로 폴더명 변경
