+++ frontend/src/shared/lib/getLineInfo.ts
