+++ frontend/src/shared/lib/numberToCharCode.ts
